+++ src/oracle-database/azext_oracle_database/tests/latest/test_oracle_database.py
@@ -40,7 +40,7 @@ class OracleDatabaseScenario(ScenarioTest):
                           '--cloudexadatainfrastructurename OFake_PowerShellExaInfra '
                           '--resource-group PowerShellTestRg ')
         self.cmd('az oracle-database cloud-vm-cluster create '
-                          '--cloud-exadata-infrastructure-id /subscriptions/fd42b73d-5f28-4a23-ae7c-ca08c625fe07/resourceGroups/PowerShellTestRg/providers/Oracle.Database/cloudExadataInfrastructures/OFake_PowerShellExaInfra '
+                          '--cloud-exadata-infrastructure-id /subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/PowerShellTestRg/providers/Oracle.Database/cloudExadataInfrastructures/OFake_PowerShellExaInfra '
                           '--resource-group PowerShellTestRg '
                           '--location eastus '
                           '--cpu-core-count 6 '
@@ -57,8 +57,8 @@ class OracleDatabaseScenario(ScenarioTest):
                           '--license-model LicenseIncluded '
                           '--memory-size-in-gbs 60 '
                           '--ssh-public-keys \'ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAABgQDHJDAu814QnRaYFbuFswJX1IdpFAa8yVk4sneI3Q+QLoTyE4W5GR5f6TR5FqA03zxp41chkvjLysnc+EYJmjXQIbvrsEfaUenGu2AbYE3szBK9v+8iUy8JLgjoGuuxTU3BNCvMlTd00yw/qrOYbEDU9ZBnZVY/0nv2E03AzGUUObZJ8IUgNRUmNdWVFQiVUcmkRRM+XsIEUpuh0S6YJLLweZd+H50Y1mhKWXnZZH8Ed/5EmkSr7cO5WEKU2O/KSbavybkjUVWi6dcQMfwmNLBH9aByqAW8QvcZkZDvxLmimLnm3Jd/QDVvGesjyLbUrWfpbsaDXs+DgjTlFlxyqLPYMON5cfSg8wBj3Y176yWxwmrPnkro8X1Y93poSDQZb9SU68DsTrgVa6FoXWPkbUXz/nZX9GkTwE1Nhy2EPL4J+J50ZUZWG0bK25dFTKrzimLn1Qmvrx3so9qDId9LWbpYI6cJYxDTkGdGpuaHqDqGAi+5HeeXxx3/zO8pErPxy20= generated-by-azure\' '
-                          '--subnet-id /subscriptions/fd42b73d-5f28-4a23-ae7c-ca08c625fe07/resourceGroups/PowerShellTestRg/providers/Microsoft.Network/virtualNetworks/PSTestVnet/subnets/delegated  '
+                          '--subnet-id /subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/PowerShellTestRg/providers/Microsoft.Network/virtualNetworks/PSTestVnet/subnets/delegated  '
                           '--time-zone UTC '
-                          '--vnet-id /subscriptions/fd42b73d-5f28-4a23-ae7c-ca08c625fe07/resourceGroups/PowerShellTestRg/providers/Microsoft.Network/virtualNetworks/PSTestVnet '
+                          '--vnet-id /subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/PowerShellTestRg/providers/Microsoft.Network/virtualNetworks/PSTestVnet '
                           '--no-wait ')