+++ src/oracle-database/azext_oracle_database/tests/latest/test_oracle_database_adbs.py
@@ -26,13 +26,13 @@ class OracleDatabaseAdbsScenario(ScenarioTest):
         self.cmd('az oracle-database autonomous-database create --location eastus '
                 '--autonomousdatabasename ADBScli '
                 '--resource-group PowerShellTestRg '
-                '--subnet-id /subscriptions/fd42b73d-5f28-4a23-ae7c-ca08c625fe07/resourceGroups/PowerShellTestRg/providers/Microsoft.Network/virtualNetworks/PSTestVnet/subnets/delegated '
+                '--subnet-id /subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/PowerShellTestRg/providers/Microsoft.Network/virtualNetworks/PSTestVnet/subnets/delegated '
                 '--display-name ADBScli1 '
                 '--compute-model ECPU --compute-count 2 '
                 '--data-storage-size-in-gbs 32 --license-model BringYourOwnLicense '
                 '--db-workload OLTP --admin-password TestPass#2024# '
                 '--db-version 19c --character-set AL32UTF8 --ncharacter-set AL16UTF16 '
-                '--vnet-id /subscriptions/fd42b73d-5f28-4a23-ae7c-ca08c625fe07/resourceGroups/PowerShellTestRg/providers/Microsoft.Network/virtualNetworks/PSTestVnet  --regular --no-wait ')
+                '--vnet-id /subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/PowerShellTestRg/providers/Microsoft.Network/virtualNetworks/PSTestVnet  --regular --no-wait ')
         self.cmd('az oracle-database autonomous-database delete '
                 '--autonomousdatabasename ADBScli '
                 '--resource-group PowerShellTestRg --yes --no-wait ')